refactor(capabilities): split render into a submodule and move its drawing kinds
Decompose `aether-capabilities/src/render.rs` (1844 lines) into a
`render/` directory along its cohesion seams — `mod` (cap struct,
`#[bridge]` dispatch, headless companion), `pipeline` (GPU bundle +
accumulator handles), `texture` (texture registry), `quad` (quad-batch
accumulator), `capture` (cross-thread readback), and `kinds` — and move
the render cap's drawing/texture mail kinds out of `aether-kinds` into
`render/kinds.rs`, the ADR-0121 "capabilities own their kinds" slice for
render.

Moved (no upstream-by-type consumer): `DrawTriangle` / `Vertex` /
`DRAW_TRIANGLE_BYTES`, `Camera`, `CreateTexture` / `CreateTextureResult`,
`UpdateTexture`, `TexturedQuad` / `DrawTexturedQuads`, `SolidQuad` /
`DrawSolidQuads`, plus their codec/name/schema tests.

Stayed central (a move would close a dependency cycle): the capture
kinds (`CaptureFrame` / `CaptureFrameResult` / `SimilarityCheck`,
consumed by `aether-mcp`), the `FrameCheck` verification family
(consumed by the substrate core), `NamedMail`, and `Mat4Apply`. Also
kept central are `QuadSpace` / `QuadScale`: `aether-kinds`' own
`aether.text.draw` kind consumes `QuadSpace`, and `aether-kinds` has no
dependency on `aether-capabilities`, so moving them would close an
`aether-kinds -> aether-capabilities` cycle — the moved quad draw kinds
import them back from `aether-kinds`.

The `render` marker feature now pulls `bytemuck` (wasm-safe) so a guest
on the marker-only feature sees the moved cast-shaped kinds, and `text`
pulls `render` because the text cap receives the render-owned
`CreateTextureResult` reply. Wire ids are unchanged (a kind id is
`fnv1a_64` over name + schema, neither touched by a crate move). Import
paths sweep `aether_kinds::` -> `aether_capabilities::render::` for the
moved kinds across the chassis bundle, kit, mesh-viewer, and fixtures.
