+++ crates/aether-mesh-viewer/src/runtime.rs
@@ -37,11 +37,10 @@ use aether_actor::{
 };
 use aether_capabilities::fs::FsMailboxExt;
 use aether_capabilities::lifecycle::LifecycleMailboxExt;
+use aether_capabilities::render::{DrawTriangle, Vertex};
 use aether_capabilities::{FsCapability, LifecycleCapability, RenderCapability};
 use aether_data::Kind;
-use aether_kinds::{
-    DrawTriangle, MeshLoadResult, ReadResult, Render, TrajectorySampleEntry, Vertex,
-};
+use aether_kinds::{MeshLoadResult, ReadResult, Render, TrajectorySampleEntry};
 use aether_labyrinth::{CorridorGraph, EdgeKind, ScalarField, TrajectorySet};
 use aether_math::Vec3;
 use aether_mesh::{Point3, Polygon, tessellate_polygon};
