+++ crates/aether-test-fixtures/bundle/src/cube.rs
@@ -31,8 +31,9 @@ use core::f32::consts::FRAC_PI_4;
 
 use aether_actor::{ActorInitError, WasmActor, WasmCtx, WasmInitCtx, actor};
 use aether_capabilities::lifecycle::LifecycleMailboxExt;
+use aether_capabilities::render::{Camera, DrawTriangle, Vertex};
 use aether_capabilities::{LifecycleCapability, RenderCapability};
-use aether_kinds::{Camera, DrawTriangle, Tick, Vertex};
+use aether_kinds::Tick;
 use aether_math::{Mat4, Vec3};
 
 /// Half-extent of the unit cube: corners sit at ±`HALF` on every axis,
