+++ crates/aether-test-fixtures/bundle/src/ui_widget.rs
@@ -25,8 +25,9 @@
 
 use aether_actor::{ActorInitError, WasmActor, WasmCtx, WasmInitCtx, actor};
 use aether_capabilities::lifecycle::LifecycleMailboxExt;
+use aether_capabilities::render::{DrawSolidQuads, SolidQuad};
 use aether_capabilities::{LifecycleCapability, RenderCapability};
-use aether_kinds::{DrawSolidQuads, QuadSpace, SolidQuad, Tick};
+use aether_kinds::{QuadSpace, Tick};
 use aether_test_fixtures_kinds::UiWidgetConfig;
 
 pub struct UiWidget {
